pulls route info , didn't parse the json cause i don't know what you nee out of it'

+++ run_me.php
@@ -107,29 +107,74 @@ class navitia_query
 			exit("not found in paris");
 	}
 
+	private function get_route_departures($route_code)
+	{
+		$qu = "https://api.navitia.io/v1/coverage/fr-idf/routes/" . $route_code . "/departures?";
+		echo $qu . "\n";
+		if (!file_exists('tmp_depart' . $route_code . '.json'))
+		{
+			$cu = curl_init();
+			curl_setopt($cu, CURLOPT_URL, $qu);
+			curl_setopt($cu, CURLOPT_USERPWD, $this->auth_token);
+			curl_setopt($cu, CURLOPT_RETURNTRANSFER, true);
+			$out = curl_exec($cu);
+			curl_close($cu);
+			file_put_contents('tmp_depart' . $route_code . '.json', $out);
+		}
+		else
+			echo $route_code . 'route.json already exists';
+		return ;
+	}
+
+	private function get_route_arrival($route_code)
+	{
+		$qu = "https://api.navitia.io/v1/coverage/fr-idf/routes/" . $route_code . "/arrivals?";
+		echo $qu . "\n";
+		if (!file_exists('tmp_arrival' . $route_code . '.json'))
+		{
+			$cu = curl_init();
+			curl_setopt($cu, CURLOPT_URL, $qu);
+			curl_setopt($cu, CURLOPT_USERPWD, $this->auth_token);
+			curl_setopt($cu, CURLOPT_RETURNTRANSFER, true);
+			$out = curl_exec($cu);
+			curl_close($cu);
+			file_put_contents('tmp_arrival' . $route_code . '.json', $out);
+		}
+		else
+			echo $route_code . 'route.json already exists';
+		return ;
+	}
+
 	public function get_all_routes()
 	{
-		$qu = "https://api.navitia.io/v1/coverage/fr-idf/networks/network:0:439/routes?depth=3&disable_geojson=true";
+		$qu = "https://api.navitia.io/v1/coverage/fr-idf/networks/network:0:439/routes?depth=3&disable_geojson=true&count=50";
 		echo $qu . "\n";
-		$cu = curl_init();
-		curl_setopt($cu, CURLOPT_URL, $qu);
-		curl_setopt($cu, CURLOPT_USERPWD, $this->auth_token);
-		curl_setopt($cu, CURLOPT_RETURNTRANSFER, true);
-		$out = curl_exec($cu);
-		curl_close($cu);
-		$tmp_arr = json_decode($out, true);
-		$ret = $tmp_arr['routes'];
-		$my_route_stop_points = array();
-		foreach ($ret as $i)
+		if (!file_exists('tmp_all_rout_info.json'))
 		{
-			foreach ($i['stop_points'] as $stop)
+			$cu = curl_init();
+			curl_setopt($cu, CURLOPT_URL, $qu);
+			curl_setopt($cu, CURLOPT_USERPWD, $this->auth_token);
+			curl_setopt($cu, CURLOPT_RETURNTRANSFER, true);
+			$out = curl_exec($cu);
+			curl_close($cu);
+			$tmp_arr = json_decode($out, true);
+			$ret = $tmp_arr['routes'];
+			$my_route_stop_points = array();
+			foreach ($ret as $i)
 			{
-				$my_route_stop_points[$i['id']][] = array('label' => $stop['stop_area']['label'],
-					'id' => $stop['id']);
+				foreach ($i['stop_points'] as $stop)
+				{
+					$my_route_stop_points[$i['id']][] = array('label' => $stop['stop_area']['label'],
+						'id' => $stop['id']);
+				}
+				$this->get_route_departures($i['id']);
+				$this->get_route_arrival($i['id']);
 			}
+			file_put_contents('tmp_semi_parsed_stop_points.json', json_encode($my_route_stop_points));
+			file_put_contents('tmp_all_rout_info.json', json_encode($ret));
 		}
-		file_put_contents('tmp_semi_parsed_stop_points.json', json_encode($my_route_stop_points));
-		file_put_contents('tmp_all_rout_info.json', json_encode($ret));
+		else
+			echo 'the route json already exists';
 	}
 
 	#get metro station
@@ -166,9 +211,17 @@ if ($len == 3)
 	#	print_r($station_final);
 		#
 		##
-		$cl->get_all_routes("");
-		file_put_contents('tmp_start.json', json_encode($station_start));
-		file_put_contents('tmp_final.json', json_encode($station_final));
+	#	if (file_exists("tmp_semi_parsed_stop_points.json"))
+		if (0)
+			;
+		else
+		{
+			$cl->get_all_routes("");
+			file_put_contents('tmp_start.json', json_encode($station_start));
+			file_put_contents('tmp_final.json', json_encode($station_final));
+		}
+#		$cl->get_network_arivals();
+#		$cl->get_network_departures();
 	}
 	else
 		echo "please input two valid strings";
